refactor(mcp): extract live control and sse relay helpers
