+++ apps/desktop/src-tauri/src/external_task_sync.rs
@@ -1,8 +1,8 @@
+use crate::sse_relay::{build_sse_stream_request, read_next_sse_message, SseEventId};
 use anyhow::{Context, Result};
 use host_application::AppService;
 use host_domain::now_rfc3339;
 use reqwest::blocking::{Client, Response};
-use reqwest::header::{HeaderMap, HeaderName, HeaderValue, ACCEPT};
 use serde::{Deserialize, Serialize};
 use std::io::{BufRead, BufReader};
 use std::sync::atomic::{AtomicBool, Ordering};
@@ -14,7 +14,6 @@ use uuid::Uuid;
 const TASK_EVENT_STREAM_PATH: &str = "task-events";
 const TASK_EVENT_RELAY_RETRY_DELAY: Duration = Duration::from_secs(1);
 const TASK_EVENT_RELAY_CONNECT_TIMEOUT: Duration = Duration::from_secs(5);
-const LAST_EVENT_ID_HEADER: HeaderName = HeaderName::from_static("last-event-id");
 pub(crate) const TASK_EVENT_NAME: &str = "openducktor://task-event";
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
@@ -35,12 +34,6 @@ pub(crate) struct ExternalTaskSyncEvent {
 
 pub(crate) struct TaskEventRelayState {
     pub(crate) stop_requested: Arc<AtomicBool>,
-}
-
-#[derive(Debug, Default, PartialEq, Eq)]
-struct SseMessage {
-    id: Option<u64>,
-    data: String,
 }
 
 pub(crate) fn build_external_task_created_event(
@@ -86,7 +79,7 @@ fn run_task_event_relay<R: tauri::Runtime>(
         .build()
         .context("failed to build task event relay HTTP client")?;
 
-    let mut last_event_id = None;
+    let mut last_event_id: Option<SseEventId> = None;
     while !stop_requested.load(Ordering::SeqCst) {
         let stream_url = match service.mcp_bridge_base_url() {
             Ok(base_url) => format!("{base_url}/{TASK_EVENT_STREAM_PATH}"),
@@ -101,20 +94,19 @@ fn run_task_event_relay<R: tauri::Runtime>(
             }
         };
 
-        let response =
-            match build_task_event_stream_request(&client, &stream_url, last_event_id).send() {
-                Ok(response) => response,
-                Err(error) => {
-                    tracing::error!(
-                        target: "openducktor.task-sync",
-                        stream_url,
-                        error = %error,
-                        "Task event relay failed to connect to the MCP bridge task stream"
-                    );
-                    sleep_before_retry(&stop_requested);
-                    continue;
-                }
-            };
+        let response = match build_sse_stream_request(&client, &stream_url, last_event_id).send() {
+            Ok(response) => response,
+            Err(error) => {
+                tracing::error!(
+                    target: "openducktor.task-sync",
+                    stream_url,
+                    error = %error,
+                    "Task event relay failed to connect to the MCP bridge task stream"
+                );
+                sleep_before_retry(&stop_requested);
+                continue;
+            }
+        };
 
         if let Err(error) =
             relay_task_event_stream(&app, response, &stop_requested, &mut last_event_id)
@@ -137,7 +129,7 @@ fn relay_task_event_stream<R: tauri::Runtime>(
     app: &AppHandle<R>,
     response: Response,
     stop_requested: &AtomicBool,
-    last_event_id: &mut Option<u64>,
+    last_event_id: &mut Option<SseEventId>,
 ) -> Result<()> {
     let status = response.status();
     if !status.is_success() {
@@ -163,7 +155,7 @@ fn relay_task_event_stream<R: tauri::Runtime>(
 fn relay_task_event_reader(
     mut reader: impl BufRead,
     stop_requested: &AtomicBool,
-    last_event_id: &mut Option<u64>,
+    last_event_id: &mut Option<SseEventId>,
     mut emit: impl FnMut(ExternalTaskSyncEvent),
 ) -> Result<()> {
     while !stop_requested.load(Ordering::SeqCst) {
@@ -196,68 +188,6 @@ fn relay_task_event_reader(
     Ok(())
 }
 
-fn read_next_sse_message(reader: &mut impl BufRead) -> Result<Option<SseMessage>> {
-    let mut message = SseMessage::default();
-
-    let mut payload_lines = Vec::new();
-
-    loop {
-        let mut line = String::new();
-        let bytes_read = reader.read_line(&mut line)?;
-
-        if bytes_read == 0 {
-            if payload_lines.is_empty() {
-                return Ok(None);
-            }
-
-            message.data = payload_lines.join("\n");
-            return Ok(Some(message));
-        }
-
-        let trimmed = line.trim_end_matches(['\r', '\n']);
-        if trimmed.is_empty() {
-            if payload_lines.is_empty() {
-                continue;
-            }
-
-            message.data = payload_lines.join("\n");
-            return Ok(Some(message));
-        }
-
-        if let Some(id) = trimmed.strip_prefix("id:") {
-            let parsed_id = id.trim().parse::<u64>().with_context(|| {
-                format!(
-                    "invalid task event stream id received from MCP bridge: {}",
-                    id.trim()
-                )
-            })?;
-            message.id = Some(parsed_id);
-            continue;
-        }
-
-        if let Some(payload) = trimmed.strip_prefix("data:") {
-            payload_lines.push(payload.trim_start().to_string());
-        }
-    }
-}
-
-fn build_task_event_stream_request<'a>(
-    client: &'a Client,
-    stream_url: &'a str,
-    last_event_id: Option<u64>,
-) -> reqwest::blocking::RequestBuilder {
-    let mut headers = HeaderMap::new();
-    headers.insert(ACCEPT, HeaderValue::from_static("text/event-stream"));
-
-    if let Some(last_event_id) = last_event_id {
-        if let Ok(value) = HeaderValue::from_str(&last_event_id.to_string()) {
-            headers.insert(LAST_EVENT_ID_HEADER.clone(), value);
-        }
-    }
-
-    client.get(stream_url).headers(headers)
-}
-
 fn sleep_before_retry(stop_requested: &AtomicBool) {
     if stop_requested.load(Ordering::SeqCst) {
         return;
@@ -280,50 +210,6 @@ mod tests {
         assert_eq!(event.task_id, "task-1");
         assert!(!event.event_id.is_empty());
         assert!(!event.emitted_at.is_empty());
-    }
-
-    #[test]
-    fn read_next_sse_message_collects_multiline_data_events_and_event_id() {
-        let mut reader = Cursor::new(
-            b"id: 7\nevent: message\ndata: {\"one\":1}\ndata: {\"two\":2}\n\n".to_vec(),
-        );
-
-        let message = read_next_sse_message(&mut reader)
-            .expect("payload should parse")
-            .expect("payload should be present");
-
-        assert_eq!(message.id, Some(7));
-        assert_eq!(message.data, "{\"one\":1}\n{\"two\":2}");
-    }
-
-    #[test]
-    fn read_next_sse_message_skips_keepalive_comments() {
-        let mut reader = Cursor::new(b": keepalive\n\ndata: {\"ok\":true}\n\n".to_vec());
-
-        let message = read_next_sse_message(&mut reader)
-            .expect("payload should parse")
-            .expect("payload should be present");
-
-        assert_eq!(message.id, None);
-        assert_eq!(message.data, "{\"ok\":true}");
-    }
-
-    #[test]
-    fn build_task_event_stream_request_includes_last_event_id_for_resume() {
-        let client = Client::builder().build().expect("client should build");
-        let request =
-            build_task_event_stream_request(&client, "http://127.0.0.1:1234/task-events", Some(42))
-                .build()
-                .expect("request should build");
-
-        assert_eq!(
-            request.headers().get(LAST_EVENT_ID_HEADER.clone()),
-            Some(&HeaderValue::from_static("42"))
-        );
-        assert_eq!(
-            request.headers().get(ACCEPT),
-            Some(&HeaderValue::from_static("text/event-stream"))
-        );
     }
 
     #[test]
@@ -351,7 +237,7 @@ mod tests {
 
         assert_eq!(last_event_id, Some(5));
 
-        let request = build_task_event_stream_request(
+        let request = build_sse_stream_request(
             &Client::builder().build().expect("client should build"),
             "http://127.0.0.1:1234/task-events",
             last_event_id,
@@ -359,8 +245,8 @@ mod tests {
         .build()
         .expect("request should build");
         assert_eq!(
-            request.headers().get(LAST_EVENT_ID_HEADER.clone()),
-            Some(&HeaderValue::from_static("5"))
+            request.headers().get("last-event-id"),
+            Some(&"5".parse().expect("header should parse"))
         );
 
         let replay_stream = Cursor::new(format!("id: 6\ndata: {second_event}\n\n").into_bytes());
@@ -397,7 +283,7 @@ mod tests {
         assert_eq!(last_event_id, Some(5));
         assert!(emitted_task_ids.is_empty());
 
-        let request = build_task_event_stream_request(
+        let request = build_sse_stream_request(
             &Client::builder().build().expect("client should build"),
             "http://127.0.0.1:1234/task-events",
             last_event_id,
@@ -405,8 +291,8 @@ mod tests {
         .build()
         .expect("request should build");
         assert_eq!(
-            request.headers().get(LAST_EVENT_ID_HEADER.clone()),
-            Some(&HeaderValue::from_static("5"))
+            request.headers().get("last-event-id"),
+            Some(&"5".parse().expect("header should parse"))
         );
     }
 }
